+++ kotlin/lib/src/main/kotlin/Statistics.kt
@@ -1,21 +1,16 @@
 // this file is @generated
 package com.svix.kotlin
 
-import com.svix.kotlin.exceptions.ApiException
-import com.svix.kotlin.internal.apis.StatisticsApi
 import com.svix.kotlin.models.AggregateEventTypesOut
 import com.svix.kotlin.models.AppUsageStatsIn
 import com.svix.kotlin.models.AppUsageStatsOut
+import okhttp3.Headers
+import okhttp3.HttpUrl
 
-class Statistics internal constructor(token: String, options: SvixOptions) {
-    private val api = StatisticsApi(options.serverUrl)
+data class StatisticsAggregateAppStatsOptions(val idempotencyKey: String? = null)
 
-    init {
-        api.accessToken = token
-        api.userAgent = options.getUA()
-        options.initialRetryDelayMillis?.let { api.initialRetryDelayMillis = it }
-        options.numRetries?.let { api.numRetries = it }
-    }
+class Statistics(baseUrl: HttpUrl, defaultHeaders: Map<String, String>) :
+    SvixHttpClient(baseUrl, defaultHeaders) {
 
     /**
      * Creates a background task to calculate the message destinations for all applications in the
@@ -26,13 +21,18 @@ class Statistics internal constructor(token: String, options: SvixOptions) {
      */
     suspend fun aggregateAppStats(
         appUsageStatsIn: AppUsageStatsIn,
-        options: PostOptions = PostOptions(),
+        options: StatisticsAggregateAppStatsOptions = StatisticsAggregateAppStatsOptions(),
     ): AppUsageStatsOut {
-        try {
-            return api.v1StatisticsAggregateAppStats(appUsageStatsIn, options.idempotencyKey)
-        } catch (e: Exception) {
-            throw ApiException.wrap(e)
-        }
+        val url = this.newUrlBuilder().encodedPath("/api/v1/stats/usage/app")
+        var headers = Headers.Builder()
+        options.idempotencyKey?.let { headers = headers.add("idempotency-key", it) }
+
+        return this.executeRequest<AppUsageStatsIn, AppUsageStatsOut>(
+            "POST",
+            url.build(),
+            headers = headers.build(),
+            reqBody = appUsageStatsIn,
+        )
     }
 
     /**
@@ -43,10 +43,7 @@ class Statistics internal constructor(token: String, options: SvixOptions) {
      * endpoint to retrieve the results of the operation.
      */
     suspend fun aggregateEventTypes(): AggregateEventTypesOut {
-        try {
-            return api.v1StatisticsAggregateEventTypes()
-        } catch (e: Exception) {
-            throw ApiException.wrap(e)
-        }
+        val url = this.newUrlBuilder().encodedPath("/api/v1/stats/usage/event-types")
+        return this.executeRequest<Any, AggregateEventTypesOut>("PUT", url.build())
     }
 }